Update msgbox
A simple update so message boxes can now also use titles for a bit more cleaner and organized look when desired. For example if creating a multi menu/message script.

This update if fully backwards compatible with existing scripts because if the second parameter isn't given, then the old original code will be executed of displaying a message box (dialog) without title.